+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 2014-06-04
 * Update setter methods in [CreatedRequest](https://github.com/jlinn/stripe-api-java/blob/master/src%2Fmain%2Fjava%2Fnet%2Fjoelinn%2Fstripe%2Frequest%2FCreatedRequest.java)
+* Add statementDescription to [InvoiceResponse](https://github.com/jlinn/stripe-api-java/blob/master/src%2Fmain%2Fjava%2Fnet%2Fjoelinn%2Fstripe%2Fresponse%2Finvoices%2FInvoiceResponse.java)
 
 ### 2014-05-29
 * Clean up [Client](https://github.com/jlinn/stripe-api-java/blob/master/src%2Fmain%2Fjava%2Fnet%2Fjoelinn%2Fstripe%2FClient.java)